// src/repositories_v2/market/types.ts
export type ParsedMarketPoolData = {
  coinType: string;
  maxBorrowRate: number;
  borrowRate: number;
  borrowRateScale: number;
  borrowIndex: number;
  lastUpdated: number;
  cashAmount: number;
  debtAmount: number;
  marketCoinSupplyAmount: number;
  reserveAmount: number;
  reserveFactor: number;
  borrowWeight: number;
  borrowFee: number;
  baseBorrowRate: number;
  borrowRateOnHighKink: number;
  borrowRateOnMidKink: number;
  highKink: number;
  midKink: number;
  minBorrowAmount: number;
  isIsolated: boolean;
  supplyLimit: number;
  borrowLimit: number;
};
// ...
export type OriginMarketCollateralData = {
  type: string;
  isIsolated: boolean;
  collateralFactor: { value: string };
  liquidationFactor: { value: string };
  liquidationDiscount: { value: string };
  liquidationPenalty: { value: string };
  liquidationReserveFactor: { value: string };
  maxCollateralAmount: string;
  totalCollateralAmount: string;
};

export type ParsedMarketCollateralData = {
  coinType: string;
  collateralFactor: number;
  liquidationFactor: number;
  liquidationDiscount: number;
  liquidationPenalty: number;
  liquidationReserveFactor: number;
  maxCollateralAmount: number;
  totalCollateralAmount: number;
  isIsolated: boolean;
};